*Added Completedobject and Totobjects counters *Started StrataRunner *Started StrataImage *Started StrataArchive

+++ Source/Configuration/StrataConfig.hpp
@@ -229,8 +229,8 @@ class StrataConfig
         xmlXPathContextPtr configXPathContext;
 
     private:
-        int completeObjects;
-        int totalObjects;
+        unsigned long completeObjects;
+        unsigned long totalObjects;
         bool configLoaded;
     
         //Is the GameMediaSource a expansion disc?
